banana nanocube fix

+++ src/main/java/vmarcinko/nanocube/Node2.java
@@ -3,7 +3,7 @@ package vmarcinko.nanocube;
 import java.util.List;
 
 public interface Node2 {
-    int getLabel();
+    long getLabel();
 
     Node2 addChild(int label);
